+++ htdocs/bimpcore/classes/components/BC_Field.php
@@ -150,7 +150,6 @@ class BC_Field extends BimpComponent
         }
 
         if ($this->edit) {
-
             if ($this->params['editable'] && $this->object->canEditField($this->name) && $this->object->isFieldEditable($this->name, $this->force_edit)) {
                 $html .= $this->renderInput();
             } else {
@@ -160,7 +159,6 @@ class BC_Field extends BimpComponent
         } else {
             $html .= $this->displayValue();
         }
-
 
         $current_bc = $prev_bc;
         return $html;